Code for the Water Jug Problem
Question:
You are given two jugs, a m-gallon one and a n-gallon
one, a pump which has unlimited water which you can use to fill the jug, and
the ground on which water may be poured. Neither jug has any measuring
markings on it. How can you get exactly x gallons of water in the n(or m)-gallon jug?